+++ src/models/index.ts
@@ -1,2 +1,3 @@
 export * from "./custody-error.js"
 export * from "./custody-types.js"
+export type * from "./ledger-ids.js"